refactor: reusable types
